Implements database client and updates bug client
Adds the new database client which will manage interactions with the database.
It will clear any old bugs from today and then insert the new list of today's
bugs (all in a transaction, which will rollback if things don't go smoothly).

The bugzilla client is also updated to reflect a few new changes.  The
external bugs are now stored as a jsonb column which means we have more data
for use later, and it is queryable with SQL. ExecuteQuery now takes a sharer
id to allow for more saved searches.  It also takes a list of fields to get
specific data (like external bugs) and drops some unwanted fields (like
priority).