Cleanup, bug customization, and database changes
General cleanup and more descriptive errors.  Added doc text for snapshotting.

Adds custom SingleElemSlice and Score types to enable bugzilla to nicely
marshal and unmarshal json and match the database entry.

Changes the way that we get information for the database to be closer aligned
with what would happen in OpenShift.  Username, password, and database name
will now be grabbed from environment variables.  Hostname will now be a
command line flag that defaults to "postgresql" - which is what the service in
OpenShift will be named. Also removes the logic that was grabbing database info
from the cfg file.